# openapi.schema.yaml
openapi: 3.1.0
info:
  title: annonars
  description: Genome annotation based on Rust and RocksDB
  contact:
    name: Manuel Holtgrewe
    email: manuel.holtgrewe@bih-charite.de
  license:
    name: Apache-2.0
  version: 0.41.3
paths:
...
  /api/v1/strucvars/clinvar/query:
    get:
      tags:
      - clinvar_sv
      summary: Endpoint for querying ClinVar SV annotations.
      operationId: strucvarsClinvarQuery
      parameters:
      - name: genome_release
        in: query
        description: Genome release specification.
        required: true
        schema:
          $ref: '#/components/schemas/GenomeRelease'
      - name: chromosome
        in: query
        description: Chromosome name.
        required: true
        schema:
          type: string
      - name: start
        in: query
        description: 1-based start position.
        required: true
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: stop
        in: query
        description: 1-based stop postion.
        required: true
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: variation_types
        in: query
        description: Optionally, the variant types.
        required: false
        schema:
          type:
          - array
          - 'null'
          items:
            $ref: '#/components/schemas/ClinvarExtractedVariationType'
      - name: min_overlap
        in: query
        description: Optionally, minimal overlap.
        required: false
        schema:
          type:
          - number
          - 'null'
          format: double
      - name: page_no
        in: query
        description: Optional 1-based page number.
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int32
          minimum: 0
      - name: page_size
        in: query
        description: Optional page size.
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int32
          minimum: 0
      responses:
        '200':
          description: Clinvar strucvars information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrucvarsClinvarResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomError'
...
components:
  schemas:
...
    ClinvarAggregateClassificationSet:
      type: object
      description: |-
        Used to bundle different types of Classifications (germline,
        oncogenic, somatic clinical impact) ; Supports summary as
        well as submission details.

        NB: called "typeAggregateClassificationSet" in XSD.
      properties:
        germline_classification:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ClinvarAggregatedGermlineClassification'
            description: The aggregate germline classification.
        somatic_clinical_impact:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ClinvarAggregatedSomaticClinicalImpact'
            description: The aggregate somatic clinical impact.
        oncogenicity_classification:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ClinvarAggregatedOncogenicityClassification'
            description: The aggregate oncogenicity classification.
...
    ClinvarAggregatedGermlineClassification:
      type: object
      description: |-
        Aggregated germline classification info.

        Corresponds to ``typeAggregatedGermlineClassification`` in XSD.

        nested elements
      required:
      - review_status
      - xrefs
      - citations
      - comments
      - history_records
      - conditions
      properties:
        review_status:
          $ref: '#/components/schemas/ClinvarAggregateGermlineReviewStatus'
          description: |-
            The aggregate review status based on all germline submissions
            for this record.
        description:
          type:
          - string
          - 'null'
          description: |
            We are not providing an enumeration for the values we report
            for germline classification within the xsd. Details are in
            <https://github.com/ncbi/clinvar/ClassificationOnClinVar.md>
        explanation:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ClinvarComment'
            description: |-
              Explanation is used only when the description is 'conflicting
              data from submitters' The element summarizes the conflict.
        xrefs:
          type: array
          items:
            $ref: '#/components/schemas/ClinvarXref'
          description: List of xrefs.
        citations:
          type: array
          items:
            $ref: '#/components/schemas/ClinvarCitation'
          description: List of citations.
        comments:
          type: array
          items:
            $ref: '#/components/schemas/ClinvarComment'
          description: List of comments.
        history_records:
          type: array
          items:
            $ref: '#/components/schemas/ClinvarDescriptionHistory'
          description: History information.
        conditions:
          type: array
          items:
            $ref: '#/components/schemas/ClinvarTraitSet'
          description: List of conditions.
        date_last_evaluated:
          type:
          - string
          - 'null'
          format: date-time
          description: Date of last evaluation.
        date_created:
          type:
          - string
          - 'null'
          format: date-time
          description: Date of creation.
        most_recent_submission:
          type:
          - string
          - 'null'
          format: date-time
          description: Date of most recent submission.
        number_of_submitters:
          type:
          - integer
          - 'null'
          format: int32
          description: Number of submitters.
        number_of_submissions:
          type:
          - integer
          - 'null'
          format: int32
          description: Number of submissions.
...
    ClinvarAggregatedSomaticClinicalImpact:
      type: object
      description: |-
        Aggregated somatic clinical impact info.

        nested elements
      required:
      - review_status
      - xrefs
      - citations
      - comments
      - history_records
      - conditions
      properties:
        review_status:
          $ref: '#/components/schemas/ClinvarAggregateSomaticClinicalImpactReviewStatus'
          description: |-
            The aggregate review status based on all somatic clinical
            impact submissions for this record.
        description:
          type:
          - string
          - 'null'
          description: |-
            We are not providing an enumeration for the values we report
            for somatic clinical impact classification within the xsd. Details are in
            <https://github.com/ncbi/clinvar/ClassificationOnClinVar.md>
        xrefs:
          type: array
          items:
            $ref: '#/components/schemas/ClinvarXref'
          description: List of xrefs.
        citations:
          type: array
          items:
            $ref: '#/components/schemas/ClinvarCitation'
          description: List of citations.
        comments:
          type: array
          items:
            $ref: '#/components/schemas/ClinvarComment'
          description: List of comments.
        history_records:
          type: array
          items:
            $ref: '#/components/schemas/ClinvarDescriptionHistory'
          description: History information.
        conditions:
          type: array
          items:
            $ref: '#/components/schemas/ClinvarTraitSet'
          description: List of conditions.
        date_last_evaluated:
          type:
          - string
          - 'null'
          format: date-time
          description: Date of last evaluation.
        date_created:
          type:
          - string
          - 'null'
          format: date-time
          description: Date of creation.
        most_recent_submission:
          type:
          - string
          - 'null'
          format: date-time
          description: Date of most recent submission.
        number_of_submitters:
          type:
          - integer
          - 'null'
          format: int32
          description: Number of submitters.
        number_of_submissions:
          type:
          - integer
          - 'null'
          format: int32
          description: Number of submissions.
...
    ClinvarClinicalAssertion:
      type: object
      description: |-
        A clinical assertion as submitted (SCV record).

        Corresponds to `MeasureTraitType` in XSD and `<ClinicalAssertion>` in XML
      required:
      - additional_submitters
      - record_status
      - replaces
      - replaceds
      - assertion
      - attributes
      - observed_ins
      - citations
      - comments
      - submission_names
      properties:
        clinvar_submission_id:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ClinvarSubmissionId'
            description: The ClinVar submission ID.
        clinvar_accession:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ClinvarAccession'
            description: The ClinVar SCV accessions.
        additional_submitters:
          type: array
          items:
            $ref: '#/components/schemas/ClinvarSubmitter'
          description: Optional list of additional submitters.
        record_status:
          $ref: '#/components/schemas/ClinvarClinicalAssertionRecordStatus'
          description: Record status.
        replaces:
          type: array
          items:
            type: string
          description: Replaces; mutually exclusive with replaceds
        replaceds:
          type: array
          items:
            $ref: '#/components/schemas/ClinvarClinicalAssertionRecordHistory'
          description: Replaced list; mutually exclusive with replaces
        classifications:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ClinvarClassificationScv'
            description: SCV classification.
        assertion:
          $ref: '#/components/schemas/ClinvarAssertion'
          description: The assertion.
        attributes:
          type: array
          items:
            $ref: '#/components/schemas/ClinvarClinicalAssertionAttributeSetElement'
          description: Attributes.
        observed_ins:
          type: array
          items:
            $ref: '#/components/schemas/ClinvarObservedIn'
          description: Observed in.
        simple_allele:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ClinvarAlleleScv'
            description: Allele in SCV; mutually exclusive with haplotype/genotype.
        haplotype:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ClinvarHaplotypeScv'
            description: Haplotype in SCV; mutually exclusive with allele/genotype.
        genotype:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ClinvarGenotypeScv'
            description: Genotype in SCV; mutually exclusive with allele/haplotype.
        trait_set:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ClinvarTraitSet'
            description: The trait set.
        citations:
          type: array
          items:
            $ref: '#/components/schemas/ClinvarCitation'
          description: List of citations.
        study_name:
          type:
          - string
          - 'null'
          description: Optional study name.
        study_description:
          type:
          - string
          - 'null'
          description: Optional study description.
        comments:
          type: array
          items:
            $ref: '#/components/schemas/ClinvarComment'
          description: List of comments.
        submission_names:
          type: array
          items:
            type: string
          description: List of submissions.
        date_created:
          type:
          - string
          - 'null'
          format: date-time
          description: Date of creation.
        date_last_updated:
          type:
          - string
          - 'null'
          format: date-time
          description: Date of creation.
        submission_date:
          type:
          - string
          - 'null'
          format: date-time
          description: Date of creation.
        id:
          type:
          - integer
          - 'null'
          format: int64
          description: ID.
          minimum: 0
        fda_recognized_database:
          type:
          - boolean
          - 'null'
          description: Whether it is an FDA recognized database.
...
    ClinvarClinicalAssertionRecordStatus:
      type: string
      description: Local enum for record status.
      enum:
      - current
      - replaced
      - removed
...
    ClinvarExtractedRcvRecord:
      type: object
      description: Protocol buffer for storing essential information of one RCV.
      required:
      - title
      properties:
        accession:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ClinvarVersionedAccession'
            description: The accession.
        title:
          type: string
          description: Title of RCV.
        classifications:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ClinvarRcvClassifications'
            description: Classifications (thinned out).
    ClinvarExtractedVariationType:
      type: string
      description: Enumeration for the type of the variant.
      enum:
      - insertion
      - deletion
      - snv
      - indel
      - duplication
      - tandem_duplication
      - structural_variant
      - copy_number_gain
      - copy_number_loss
      - protein_only
      - microsatellite
      - inversion
      - other
    ClinvarExtractedVcvRecord:
      type: object
      description: Protocol buffer for storing essential information of one VCV.
      required:
      - rcvs
      - name
      - variation_type
      - clinical_assertions
      - hgnc_ids
      properties:
        accession:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ClinvarVersionedAccession'
            description: The accession.
        rcvs:
          type: array
          items:
            $ref: '#/components/schemas/ClinvarExtractedRcvRecord'
          description: List of aggregated RCVs.
        name:
          type: string
          description: Name of VCV.
        variation_type:
          $ref: '#/components/schemas/ClinvarExtractedVariationType'
          description: The type of the variant.
        classifications:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ClinvarAggregateClassificationSet'
            description: Classifications (thinned out).
        clinical_assertions:
          type: array
          items:
            $ref: '#/components/schemas/ClinvarClinicalAssertion'
          description: Clinical assertions (thinned out),
        sequence_location:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ClinvarSequenceLocation'
            description: The sequence location on one reference.
        hgnc_ids:
          type: array
          items:
            type: string
          description: List of HGNC IDs.
...
    ClinvarRcvClassifications:
      type: object
      description: Local type for RCV classifications.
      properties:
        germline_classification:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ClinvarRcvGermlineClassification'
            description: Germline classification.
        somatic_clinical_impact:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ClinvarRcvAccessionSomaticClinicalImpact'
            description: Somatic clinical impact.
        oncogenicity_classification:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ClinvarRcvOncogenicityClassification'
            description: Oncogenicity classification.
...
    ClinvarSequenceLocation:
      type: object
      description: Local type for sequence location.
      required:
      - assembly
      - chr
      properties:
        for_display:
          type:
          - boolean
          - 'null'
          description: forDisplay value.
        assembly:
          type: string
          description: Name of assembly.
        chr:
          $ref: '#/components/schemas/ClinvarChromosome'
          description: Chromosomeof variant.
        accession:
          type:
          - string
          - 'null'
          description: Optional chromosome accession.
        outer_start:
          type:
          - integer
          - 'null'
          format: int32
          description: Outer start position.
          minimum: 0
        inner_start:
          type:
          - integer
          - 'null'
          format: int32
          description: Inner start position.
          minimum: 0
        start:
          type:
          - integer
          - 'null'
          format: int32
          description: Start position.
          minimum: 0
        stop:
          type:
          - integer
          - 'null'
          format: int32
          description: Stop position.
          minimum: 0
        inner_stop:
          type:
          - integer
          - 'null'
          format: int32
          description: Inner stop position.
          minimum: 0
        outer_stop:
          type:
          - integer
          - 'null'
          format: int32
          description: Outer stop position.
          minimum: 0
        display_start:
          type:
          - integer
          - 'null'
          format: int32
          description: Display start position.
          minimum: 0
        display_stop:
          type:
          - integer
          - 'null'
          format: int32
          description: Display stop position.
          minimum: 0
        strand:
          type:
          - string
          - 'null'
          description: Strand.
        variant_length:
          type:
          - integer
          - 'null'
          format: int32
          description: Variant length.
          minimum: 0
        reference_allele:
          type:
          - string
          - 'null'
          description: Reference allele.
        alternate_allele:
          type:
          - string
          - 'null'
          description: Alternate allele.
        assembly_accession_version:
          type:
          - string
          - 'null'
          description: Assembly accession version.
        assembly_status:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ClinvarAssemblyStatus'
            description: Assembly status.
        position_vcf:
          type:
          - integer
          - 'null'
          format: int32
          description: Position in VCF.
          minimum: 0
        reference_allele_vcf:
          type:
          - string
          - 'null'
          description: Reference allele in VCF.
        alternate_allele_vcf:
          type:
          - string
          - 'null'
          description: Alternate allele in VCF.
        for_display_length:
          type:
          - integer
          - 'null'
          format: int32
          description: For display length.
          minimum: 0
...
    ClinvarSubmissionId:
      type: object
      description: Local type for ClinVarSubmissionID.
      required:
      - local_key
      properties:
        local_key:
          type: string
          description: |-
            The identifier provided by the submitter to facilitate
            identification of records corresponding to their submissions. If not
            provided by a submitter, NCBI generates one. If provided by
            submitter, that is represented in localKeyIsSubmitted.
        title:
          type:
          - string
          - 'null'
          description: Optional title.
        local_key_is_submitted:
          type:
          - boolean
          - 'null'
          description: Optional indication whether local key has been submitted.
        submitted_assembly:
          type:
          - string
          - 'null'
          description: Optional assembly of submission.
...
    ClinvarVersionedAccession:
      type: object
      description: Accession with version.
      required:
      - accession
      - version
      properties:
        accession:
          type: string
          description: The accession.
        version:
          type: integer
          format: int32
          description: The version.
...
    CustomError:
      type: object
      description: Custom error type for the Actix server.
      required:
      - err
      properties:
        err:
          type: string
...
    GenomeRelease:
      type: string
      description: Local genome release for command line arguments.
      enum:
      - grch37
      - grch38
...
    StrucvarsClinvarPageInfo:
      type: object
      description: Information regarding the pagination.
      required:
      - total
      - per_page
      - current_page
      - total_pages
      properties:
        total:
          type: integer
          format: int32
          description: The total number of records.
          minimum: 0
        per_page:
          type: integer
          format: int32
          description: The number of records per page.
          minimum: 0
        current_page:
          type: integer
          format: int32
          description: The current page number.
          minimum: 0
        total_pages:
          type: integer
          format: int32
          description: The total number of pages.
          minimum: 0
...
    StrucvarsClinvarResponse:
      type: object
      description: Response of the `/api/v1/strucvars/clinvar-annos/query` endpoint.
      required:
      - records
      - page_info
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/StrucvarsClinvarResponseRecord'
          description: The records in this page.
        page_info:
          $ref: '#/components/schemas/StrucvarsClinvarPageInfo'
          description: Pagination information.
    StrucvarsClinvarResponseRecord:
      type: object
      description: Information on one response record.
      required:
      - overlap
      properties:
        record:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ClinvarExtractedVcvRecord'
            description: The record.
        overlap:
          type: number
          format: double
          description: The reciprocal overlap with the query.
...